+++ packages/livedoc-mocha/docs/ReleaseNotes.md
@@ -2,6 +2,8 @@
 
 ## 0.2.2 - beta
 * Bug: [#14](https://github.com/dotnetprofessional/LiveDoc/issues/14) - stepContext incorrect while in Background steps
+* Feature: [#15](https://github.com/dotnetprofessional/LiveDoc/issues/15) - Support Background cleanup
+* Feature: [#12](https://github.com/dotnetprofessional/LiveDoc/issues/12) - Add docStringAsEntity
 
 ## 0.2.1 - beta
 * Bug: [#10](https://github.com/dotnetprofessional/LiveDoc/issues/10) - docString on stepContext not returning full result